[Linter] Update some PHP files to simplify the phpcs rules
Summary: A couple naming convention rules are enabled by default and added to php codesniffer 3.6. So far we disabled these rule but this causes previous phpcs versions to fail reading the rule file, because they don't know that rule. This diff implement these rules which are not harmful, simplifying the rule set and making it easier to install/run the linter with no issue.

Test Plan:
With phpcs < 3.6 and >= 3.6:
  arc lint --everything

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11104