[Cashtab] Webpack upgrade part 1
Summary:
T1682

The most efficient way to upgrade Cashtab's dependency tree is to start with the latest version of Create React App. Cashtab was initially built on top of Create React App.

Because the procedure for migrating Cashtab to the latest version of Create React App involves some complications, it will be split up into steps so that each can be reviewed as a diff in phab.

See T1682 for an outline of the steps.

This diff does nothing but install the latest version of create-react-app using node 16. It does not impact any running code. It exists so that the next steps can be compared against this starting point.

Test Plan:
```
cd cashtab-v2
nvm use 16
npm i
npm start
```

Observe default create react app landing page

Reviewers: #bitcoin_abc, kieran709, emack

Reviewed By: #bitcoin_abc, kieran709, emack

Subscribers: kieran709

Differential Revision: https://reviews.bitcoinabc.org/D11120