[avalanche] Convert the remaining tests to run with the new proof format
Summary:
With the new tools from D11122, converting the remaining tests so they run with both the legacy and new proof formats is trivial.
The hardcoded proof case from  abc_rpc_addavalanchenode is removed because it no longer makes sense in this context.
The `abc_rpc_avalancheproof` test is already covering both formats and needs a separate refactor to take advantage of the new facilities.
The `abc_rpc_ecash` test uses a legacy hardcoded proof, but since it's not testing proofs at all there is no need to run with the new format.

Depends on D11122.

Test Plan:
  ninja check-functional-extended

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11123