[Cashtab] Correct proptypes validation for settings
Summary:
T2281

This prop is initialized as `false` as part of loading the parameter and validating it. This diff adjusts the proptypes validation to accept `bool` as well as `object`

Test Plan:
`npm start` from master and see react proptype error in dev console.
`npm start` from this branch and see react proptype error is gone

Reviewers: #bitcoin_abc, kieran709

Reviewed By: #bitcoin_abc, kieran709

Differential Revision: https://reviews.bitcoinabc.org/D11124