[avalanche] Add a getavalancheinfo RPC for general status info
Summary:
This returns some general informations about how many proofs we know, the total staked amount and how many avalanche nodes we are connected to. This can be used to monitor the avalanche network.

The avalanche query timeout is made configurable to let the test run in a reasonable amount of time (the default timeout is 10s).

Depends on D11131 and D11132.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, Mengerian, tyler-smith

Reviewed By: #bitcoin_abc, Mengerian, tyler-smith

Subscribers: tyler-smith, Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D11133