[Cashtab] Webpack upgrade part 2, ejecting create react app
Summary:
T1682

This diff ejects create react app following step 2 outlined in the task.

'Ejecting' is the term create-react-app uses to unpackage all of the concealed webpack configs and scripts used in the app. Ideally, we would not have to 'eject' the app, since it complicates the codebase, and these scripts are not directly related to Cashtab.

However, the latest version of create-react-app does not include some polyfill support that is required by key dependencies of Cashtab. So, the app must be ejected to allow polyfill support to be added to webpack.

This diff does not impact the production version of the app. All of these files were generated by running `npm run eject` from the `web/cashtab-v2` folder.

Test Plan:
```
cd web/cashtab-v2
nvm use 16
npm start
```

Observe create-react-app start page

Reviewers: kieran709, #bitcoin_abc

Reviewed By: kieran709, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11143