[avalanche] Fix potential abc_rpc_getavalancheinfo test failure on slow machines
Summary: The proof validation is done at each block and the test assumes that the validation occurs immediately after the node mined a block, but this assumption might be wrong on slow machines and the test has been seen failing when run with qemu. We can just wait for the output to turn right instead of asserting it immediately.

Test Plan:
On a slow machine:
  ./test/functional/test_runner.py abc_rpc_getavalancheinfo

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11163