[Cashtab] Support param parsing in QR scanned addresses
Summary: Related to task T2313. Added check for param parsing in QR scanned addresses. In the parseContent function in ScanQrCode.js, there is a check for '?', if one is present, the string will be split on the '?'  allowing the valid address to be parsed normally.

Test Plan:
cd web/cashtab && npm start
Navigate to send tab

On another device, generate a QR code that contains a valid eCash address with params added to the end

scan the QR code from the machine running this patch and observe the address field populate with the correct address

Repeat preceding two steps with a valid eToken address incl params

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11217