[Cashtab] Webpack upgrade, part 5
Summary:
T1682

Remove all the default create react code files that will not be used after the Cashtab codebase is dropped into this updated version of create-react-app.

Test Plan:
The app should not work anymore.

Review T1682 to understand the context of this step. We'll do a feature freeze after the Airdrop diff and get to a working upgrade, deprecating the old create-react-app Cashtab.

Reviewers: #bitcoin_abc, kieran709

Reviewed By: #bitcoin_abc, kieran709

Differential Revision: https://reviews.bitcoinabc.org/D11233