[avalanche] Don't poll or vote definitively until we have quorum.
Summary:
Add parameters to the Avalanche Processor that control how we decide if a quorum is available. Until a quorum is established, don't send polls and when polled always vote UNKNOWN.

Parameters added are the minimum amount of stake across the Avalanche network that must be present, and the ratio of that stake that we must have connections to.

The default parameters are set to 0 so that existing tests, written against an automatic quorum, continue to pass. These defaults and tests will be updated in a follow-on revision later.

Depends on D11181

Test Plan: ninja check-avalanche

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Maniphest Tasks: T1864

Differential Revision: https://reviews.bitcoinabc.org/D11101