[Cashtab-v2] Patch Send Confirmations text color in Configure.js
Summary: Related to task T2340. In cashtab-v2's Configure tab, the font color of Send Confirmations does not match the v1 styling. Added class selector to GeneralSettingsItem and added the color lightwhite to the text in that div.

Test Plan:
cd web/cashtab-v2 && npm start
navigate to the configure tab
observe that the font color is lightwhite, not black

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11252