[Cashtab-v2] Webpack 5 CSS patch: ant-descriptions-title
Summary: Related to task T2347. On the send token page, the ant-descriptions-title is rendering with black font in Cashtab-v2 and has been patched to set it to the correct color.

Test Plan:
cd web/cashtab-v2
npm start
navigate to send token page
observe that the Token Info for <Token> text is the correct color

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11262