[.gitignore] Add .netlify folder for webdev
Summary:
T2358

This addition was generated by running netlify deploy in the cashtab repo. This is useful for production testing of diffs for web dev projects. The folder should not be committed to the repo as its contents should differ for different developers.

Test Plan: Review the .gitignore change and ensure this doesn't conflict with any needed folders in other parts of the repo.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11269