[avalanche] Keep sending getavaaddr until the qorum is established
Summary:
The getavaaddr is conditionally sent when we failed to find a node to query. This is not working properly for the time where the quorum is not established yet as there is no attempt to select a node, altough that's the time we need them the most. This diff fixes this isse by issuing the getavaaddr request while the qorum is  not esablished.

Ref T1864.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Maniphest Tasks: T1864

Differential Revision: https://reviews.bitcoinabc.org/D11276