[Cashtab] Add disabled button component & use in Send.js
Summary: Related to task T2359. In a previous revision, the SecondaryButton styled component was changed to show the disabled symbol on hover, however since this button is used in different contexts throughout the app, this revision adds a disabled button component.

Test Plan:
cd web/cashtab-v2 && npm start
in a private browser, hover over the import wallet button on the home screen
the cursor should be a pointer
in a non-private browser
navigate to the send screen
add invalid eCash address or invalid XEC amount
over over the send button
observe the disabled symbol on the send button.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11281