[Cashtab]  CSS patch font color of alerts in Configure.js
Summary: Related to task T2341. Font color of Seed Phrase and translation alerts in Configure.js changed to black during the webpack upgrade, however the previous color is not part of the existing theme so it has been hardcoded in.

Test Plan:
cd web/cashtab-v2
npm start
navigate to the Configure tab
observe that the alerts at the top of the page have the correct font color.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11286