[avalanche] Always allow for downloading our local proof if it has been advertised
Summary:
Our local proof can be advertised by an avahello message before it gets validated upon startup, because it is waiting for a block to proceed to the proof verification.
This can also occur during IBD, and the node will return a notfound for the proofid. This phenomenon prevents the peer from requesting it again from us and can significantly slow down the proof relaying for our node. In order to prevent this, we can simply assume our proof is valid and allow for it to be downloaded with no delay. To prevent DoS'ing it still has to be announced first.

Ref T1864.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Maniphest Tasks: T1864

Differential Revision: https://reviews.bitcoinabc.org/D11288