[avalanche] Reduce the getavaaddr message  sending interval
Summary:
Just make it a bit faster to send new getavaaddr, and make the anti-DoS interval narrower so it's not in the normal range of the message.
The test needs to be adapted to get rid of an assumption that the sending interval was greater than the max delay before getting the addresses (it's still several times larger on average).

Depends on D11293.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11294