[Cashtab] Upload token icon info and airdrop alert css patch
Summary: Related to T2360. Fixed token icon upload text and airdrop alert info rendering black instead of grey.

Test Plan:
cd web/cashtab && npm start
navigate to create token form
upload an image
observe that the paperclip icon, the file name, and the 'click here' text are all rendering as grey.
navigate to the airdrop form,
expand the collapse and observe that the text in the alert is rendering as grey.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11296