[Cashtab] Center collapse headers in Send.js
Summary: First step in task T2348. The headers of the three styled collapses in Send.js should be centered, with the arrow icon staying to the left of the collapse.

Test Plan:
cd web/cashtab && npm start
navigate to the Send page
observe that in each of the 3 styled collapses, the header is centered

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11315