Improve clarity of makeseeds and increase max ipv6 ASNs
Summary:
Backport of [[ https://github.com/bitcoin/bitcoin/pull/24818 | core#24818 ]]

This patch increases the number of ipv6 seeds from 2 to 10 per ASN and
increases MIN_BLOCKS to 730000.

Most of the refactors are not big improvements in clarity but it does bring
 our code more inline with Bitcoin Core.

Test Plan:
  python3 makeseeds.py < seeds_main.txt > nodes_main.txt
  python3 makeseeds.py < seeds_test.txt > nodes_test.txt
  git diff

Diff should be more additions than removals and retain most of the old entries.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12141