[Cashtab] remove unnecessary constant keys
Summary: Related to T2382. Removed unnecessary constant keys from Config.js & Send.js while keeping the keys that are necessary for loading a collapse item in a default state.

Test Plan:
cd web/cashtab && npm start
Reply to a cashtab message
observe the advanced collapse is open when the page loads
navigate to an eToken information page and click Airdrop Token
observe that the Airdrop collapse is open when the page loads

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11333