Add an option to allow for bypassing the network group limit on outbound connections
Summary: This hidden option is intended for testing purpose only. It makes it possible to bypass the network group limitation for outbound connection, designed to prevent from connecting to several peers located in the same area and as such helps preventing eclipse attacks. This new flag comes in handy when running tests in a semi private network made of nodes located at the same place. There is no change of behavior when the flag is not set.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Differential Revision: https://reviews.bitcoinabc.org/D11340