[avalanche] Simplify the peermanager tests by deduplicating the addCoin lambdas
Summary:
These lambdas are used all over the tests, so make it a static function to maintain it in a single place and simplify the tests.
The amount and height values are updated when possible to make the code simpler by matching the default, allowing for complete removal at a later point in time.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Differential Revision: https://reviews.bitcoinabc.org/D11346