[Cashtab] Stop content from jumping in saved wallet & contact lists
Summary: Related to task T2394. On the settings screen, when the user hovers over a wallet name, contact name, or contact address, a tooltip will appear. This replaces the hover effect which caused the content on the screen to jump on hover. Other hover effects which do not cause the content to jump are still in place.

Test Plan:
cd web/cashtab && npm start
navigate to settings page
in the saved wallet list, hover over wallet names
observe the tooltip pop up with the wallet's name
in the contacts list, over over contact names
observe the tooltip pop up with the contact's name
hover over the contact address
observe the tooltip popup with the contact's address

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11354