Fix a bug in ParseMoney that prevents from parsing large numbers
Summary: With the currency unit change, large numbers are failed to parse due to the 63 bits overflow prevention not being adjusted to account for the new decimal number. This diff fixes the issue, update the tests that was not checking for the correct error/not using the correct value and make sure we can parse the max money. The related avalanche processor test is expanded a bit for better coverage.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11358