[avalanche] Make the max number of outbound connection configurable
Summary: The current default is still fairly limited, and should be increased. Since picking a default that would fit everybody is not an easy task, make it configurable so any user can override it as needed. This will also make it easier to test. There is no change in behavior by default.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11360