[avalanche] Bump our avalanche outbound connections
Summary: The right number is network dependent and as such not easy to select. The choice has been made so that it can still be manually configured (depends on D11360) and remain in par with core as the outbound/total ratio: core has 10 outbounds over 125 connections, i.e. 8%. Since we have 4096 connections, the same ratio gives ~328 outbounds. I picked 300 because I like round numbers.

Test Plan:
  ninja check-extended

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Differential Revision: https://reviews.bitcoinabc.org/D11361